# docs/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: Maricho API
  description: Backend API for the Maricho skilled work platform.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /workers/me/skills:
    post:
      summary: Register a skill (trade) for the current worker
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trade]
              properties:
                trade:
                  type: string
                  enum: [PLUMBING, ELECTRICAL, SOLAR, WELDING, BUILDING, APPLIANCE_REPAIR]
                proof_url:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Skill registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '403':
          description: Authenticated user is not a WORKER
        '409':
          description: Skill already registered for this trade
    get:
      summary: List the current worker's registered skills
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Skills for this worker, paginated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
        '403':
          description: Authenticated user is not a WORKER
  /workers/me/service-areas:
    post:
      summary: Register a service area (suburb) for the current worker
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [suburb]
              properties:
                suburb:
                  type: string
                travel_means:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Service area registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceArea'
        '403':
          description: Authenticated user is not a WORKER
        '409':
          description: Service area already registered for this suburb
        '422':
          description: Unknown suburb
    get:
      summary: List the current worker's registered service areas
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Service areas for this worker, paginated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceArea'
        '403':
          description: Authenticated user is not a WORKER
  /workers/me/standing:
    get:
      summary: View the current worker's standing
      description: >
        Standing is computed, never assigned. A worker with no standings row
        yet (nothing registered) gets the same defaults a new row would have.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Standing for this worker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Standing'
        '403':
          description: Authenticated user is not a WORKER
  /workers/me/vetting:
    get:
      summary: View the current worker's vetting evidence and computed grade
      description: >
        Progressive vetting tiers (ADR-005). grade is computed from
        accumulated evidence — never assigned directly.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Vetting status for this worker
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerVetting'
        '403':
          description: Authenticated user is not a WORKER
  /workers/me/vetting/id-photo:
    post:
      summary: Submit (or replace) ID-photo evidence
      description: >
        Gates REGISTERED -> IDENTIFIED (ADR-005). Grade is recomputed
        immediately from all accumulated evidence.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_photo_url]
              properties:
                id_photo_url:
                  type: string
      responses:
        '200':
          description: ID photo recorded, grade recomputed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerVetting'
        '403':
          description: Authenticated user is not a WORKER
  /workers/me/references:
    post:
      summary: Add a contactable reference
      description: Gates IDENTIFIED -> APPRENTICE (ADR-005).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, phone]
              properties:
                name:
                  type: string
                phone:
                  type: string
      responses:
        '201':
          description: Reference recorded, grade recomputed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerReference'
        '403':
          description: Authenticated user is not a WORKER
    get:
      summary: List the current worker's submitted references
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: References for this worker, paginated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkerReference'
        '403':
          description: Authenticated user is not a WORKER
  /workers/me/vouches:
    post:
      summary: Vouch for another worker
      description: >
        Gates APPRENTICE -> JOURNEYMAN/EXPERT (ADR-005). Only workers whose
        own current grade is JOURNEYMAN or EXPERT may vouch. Recomputes the
        *vouched-for* worker's grade, not the voucher's.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [worker_id]
              properties:
                worker_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Vouch recorded, vouched-for worker's grade recomputed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerVouch'
        '403':
          description: >-
            Authenticated user is not a WORKER, or their grade is below
            JOURNEYMAN
        '409':
          description: This voucher has already vouched for this worker
        '422':
          description: >-
            Self-vouch attempted, or worker_id does not refer to a
            registered worker
...
  /admin/payment-gateways/ecocash:
    get:
      summary: View the EcoCash gateway configuration
      description: OPS-only. api_key is never returned — only whether one is set.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current EcoCash configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcoCashConfig'
        '403':
          description: Authenticated user is not OPS
        '404':
          description: EcoCash is not configured yet
    put:
      summary: Set (or replace) the EcoCash gateway configuration
      description: >
        OPS-only. Entered/rotated from the running system, not via env
        vars (ADR-002) — credentials can change without a redeploy.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                merchant_code:
                  type: string
                  nullable: true
                api_key:
                  type: string
                  nullable: true
                base_url:
                  type: string
                  nullable: true
                is_sandbox:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Configuration saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcoCashConfig'
        '403':
          description: Authenticated user is not OPS
  /webhooks/ecocash:
    post:
      summary: Receive an EcoCash payment-confirmation callback
      description: >
        Logs the callback as a raw, append-only audit record
        (EcoCashCallbackLog) keyed by reference — not reconciled into the
        immutable ledger tables by mutation. Authenticated via a bearer
        token matching the configured api_key — a placeholder, not a
        confirmed EcoCash signature scheme (see app/ecocash.py). No
        request body validation beyond requiring a reference field.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reference]
              properties:
                reference:
                  type: string
                status:
                  type: string
                  nullable: true
                amount:
                  type: number
                  nullable: true
                msisdn:
                  type: string
                  nullable: true
      responses:
        '204':
          description: Callback logged
        '401':
          description: Missing or incorrect Authorization header
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
      description: Maximum rows to return.
    offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Rows to skip before collecting the result set.
  schemas:
...
    Skill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        worker_id:
          type: string
          format: uuid
        trade:
          type: string
          enum: [PLUMBING, ELECTRICAL, SOLAR, WELDING, BUILDING, APPLIANCE_REPAIR]
        grade:
          type: string
          nullable: true
        proof_url:
          type: string
          nullable: true
    ServiceArea:
      type: object
      properties:
        id:
          type: string
          format: uuid
        worker_id:
          type: string
          format: uuid
        suburb:
          type: string
        travel_means:
          type: string
          nullable: true
    Standing:
      type: object
      properties:
        worker_id:
          type: string
          format: uuid
        grade:
          type: string
        on_time_rate:
          type: number
        dispute_rate:
          type: number
        fill_rate:
          type: number
        jobs_completed:
          type: integer
        is_frozen:
          type: boolean
          description: >
            True while this worker has any job currently under an open
            dispute. Computed, not stored — never dropped, per the
            "worker's worst day" design principle.
    WorkerVetting:
      type: object
      description: Progressive vetting evidence and computed grade (ADR-005).
      properties:
        worker_id:
          type: string
          format: uuid
        id_photo_url:
          type: string
          nullable: true
        id_submitted_at:
          type: string
          format: date-time
          nullable: true
        reference_count:
          type: integer
        vouch_count:
          type: integer
          description: Total vouches received, regardless of voucher grade validity.
        grade:
          type: string
    WorkerReference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        worker_id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        created_at:
          type: string
          format: date-time
    WorkerVouch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        worker_id:
          type: string
          format: uuid
          description: The worker being vouched for.
        voucher_id:
          type: string
          format: uuid
          description: The vouching worker.
        created_at:
          type: string
          format: date-time
...
    EcoCashConfig:
      type: object
      properties:
        provider:
          type: string
        merchant_code:
          type: string
          nullable: true
        base_url:
          type: string
          nullable: true
        is_sandbox:
          type: boolean
        api_key_configured:
          type: boolean
          description: Whether an api_key is set — the key itself is never returned.
        updated_at:
          type: string
          format: date-time
          nullable: true